requirements.txt
numpy~=1.23.4
pandas~=1.5.1
astropy~=5.1
parso~=0.8.3
matplotlib~=3.5.3
lmfit~=1.1.0
scipy~=1.9.3
joblib~=1.2.0
pylatex~=1.4.1
streamlit~=1.15.2
mpld3~=0.5.8
lime-stable~=0.9.11
streamlit-authenticator~=0.2.1
pillow>=9.3.0
cryptography~='37.0.4'